+++ resources/views/livewire/admin/panel/index-panel.blade.php
@@ -20,45 +20,46 @@
     <!-- Main content -->
     <section class="content">
         <div class="container-fluid">
-            <div class="row">
-                <div class="col-4 mb-3">
-                    <a href="{{ route('panel.create') }}" class="btn btn-app bg-primary">
-                        <i class="fas fa-plus"></i> Add Data
-                    </a>
-                </div>
-            </div>
-
-            <div class="card card-primary">
-                <div class="row p-3">
-                    <div class="col">
-                        @if (session()->has('message'))
-                            <div class="alert alert-success text-center">{{ session('message') }}</div>
-                        @endif
-                        <form wire:submit.prevent="addPanel">
-                            <div class="form-group">
-                                <label for="filePanel" class="font-weight-bold">Import Data Panel</label>
-                                <input type="file" class="form-control" wire:model="filePanel" />
-                            </div>
-
-                            @error('filePanel')
-                                <span class="text-danger" style="font-size: 11.5px;">{{ $message }}</span>
-                            @enderror
-
-                            <div wire:loading wire:target="filePanel" wire:key="filePanel"><i
-                                    class="fa fa-spinner fa-spin mt-2 ml-2"></i> Uploading
-                            </div>
-
-                            <div class="form-group">
-                                <button type="submit" class="btn btn-primary w-20 mt-2">
-                                    <div wire:loading wire:target="addJalan" wire:key="addJalan"><i
-                                            class="fa fa-spinner fa-spin"></i></div> Import
-                                </button>
-                            </div>
-                        </form>
+            @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                <div class="row">
+                    <div class="col-4 mb-3">
+                        <a href="{{ route('panel.create') }}" class="btn btn-app bg-primary">
+                            <i class="fas fa-plus"></i> Add Data
+                        </a>
                     </div>
                 </div>
-            </div>
 
+                <div class="card card-primary">
+                    <div class="row p-3">
+                        <div class="col">
+                            @if (session()->has('message'))
+                                <div class="alert alert-success text-center">{{ session('message') }}</div>
+                            @endif
+                            <form wire:submit.prevent="addPanel">
+                                <div class="form-group">
+                                    <label for="filePanel" class="font-weight-bold">Import Data Panel</label>
+                                    <input type="file" class="form-control" wire:model="filePanel" />
+                                </div>
+
+                                @error('filePanel')
+                                    <span class="text-danger" style="font-size: 11.5px;">{{ $message }}</span>
+                                @enderror
+
+                                <div wire:loading wire:target="filePanel" wire:key="filePanel"><i
+                                        class="fa fa-spinner fa-spin mt-2 ml-2"></i> Uploading
+                                </div>
+
+                                <div class="form-group">
+                                    <button type="submit" class="btn btn-primary w-20 mt-2">
+                                        <div wire:loading wire:target="addJalan" wire:key="addJalan"><i
+                                                class="fa fa-spinner fa-spin"></i></div> Import
+                                    </button>
+                                </div>
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            @endif
             <div class="row">
                 <div class="col-12">
                     <div class="card">
@@ -92,7 +93,9 @@
                                         <th>Saklar</th>
                                         <th>Kordinat</th>
                                         <th>Riwayat</th>
-                                        <th>Action</th>
+                                        @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                                            <th>Action</th>
+                                        @endif
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -128,17 +131,19 @@
                                             <td>
                                                 {{ $panel->riwayat->count() }}
                                             </td>
-                                            <td>
-                                                <a href="panel/{{ $panel->id }}/edit"
-                                                    class="btn btn-sm btn-primary mx-2">
-                                                    <i class="fas fa-edit"></i>
-                                                </a>
-                                                <button wire:click="deletePanel({{ $panel->id }})"
-                                                    wire:confirm="Are you sure you want to delete this Panel?"
-                                                    class="btn btn-sm btn-danger">
-                                                    <i class="fas fa-trash"></i>
-                                                </button>
-                                            </td>
+                                            @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                                                <td>
+                                                    <a href="panel/{{ $panel->id }}/edit"
+                                                        class="btn btn-sm btn-primary mx-2">
+                                                        <i class="fas fa-edit"></i>
+                                                    </a>
+                                                    <button wire:click="deletePanel({{ $panel->id }})"
+                                                        wire:confirm="Are you sure you want to delete this Panel?"
+                                                        class="btn btn-sm btn-danger">
+                                                        <i class="fas fa-trash"></i>
+                                                    </button>
+                                                </td>
+                                            @endif
                                         </tr>
                                     @endforeach
                                 </tbody>
