+++ resources/views/livewire/admin/riwayat-panel/index-riwayat-panel.blade.php
@@ -21,13 +21,15 @@
     <!-- Main content -->
     <section class="content">
         <div class="container-fluid">
-            <div class="row mb-3">
-                <div class="col-4">
-                    <a href="{{ route('riwayatPanel.create') }}" class="btn btn-app bg-primary">
-                        <i class="fas fa-plus"></i> Add Data
-                    </a>
+            @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                <div class="row mb-3">
+                    <div class="col-4">
+                        <a href="{{ route('riwayatPanel.create') }}" class="btn btn-app bg-primary">
+                            <i class="fas fa-plus"></i> Add Data
+                        </a>
+                    </div>
                 </div>
-            </div>
+            @endif
 
             <!-- /.row -->
             <div class="row">
@@ -63,7 +65,9 @@
                                         <th>Bahan</th>
                                         <th>Keterangan</th>
                                         <th>Regu</th>
-                                        <th>Action</th>
+                                        @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                                            <th>Action</th>
+                                        @endif
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -99,17 +103,19 @@
                                             <td>
                                                 {{ $riwayatPanel->regu->nama }}
                                             </td>
-                                            <td>
-                                                <a href="riwayat_panel/{{ $riwayatPanel->id }}/edit"
-                                                    class="btn btn-sm btn-primary mx-2">
-                                                    <i class="fas fa-edit"></i>
-                                                </a>
-                                                <button wire:click="deleteRiwayatPanel({{ $riwayatPanel->id }})"
-                                                    wire:confirm="Are you sure you want to delete this Riwayat?"
-                                                    class="btn btn-sm btn-danger">
-                                                    <i class="fas fa-trash"></i>
-                                                </button>
-                                            </td>
+                                            @if (Auth::user()->role == 'Admin' || Auth::user()->role == 'User')
+                                                <td>
+                                                    <a href="riwayat_panel/{{ $riwayatPanel->id }}/edit"
+                                                        class="btn btn-sm btn-primary mx-2">
+                                                        <i class="fas fa-edit"></i>
+                                                    </a>
+                                                    <button wire:click="deleteRiwayatPanel({{ $riwayatPanel->id }})"
+                                                        wire:confirm="Are you sure you want to delete this Riwayat?"
+                                                        class="btn btn-sm btn-danger">
+                                                        <i class="fas fa-trash"></i>
+                                                    </button>
+                                                </td>
+                                            @endif
                                         </tr>
                                     @endforeach
                                 </tbody>
